Added delete function to employee dataprovider. renamed misspelled file and renamed file for clarity
