+++ src/components/employee/EmployeeProvider.js
@@ -22,9 +22,16 @@ export const EmployeeProvider = (props) => {
         .then(getEmployees)
     }
 
+    const removeEmployee = employeeId =>{
+        return fetch(`http://localhost:8088/employees/${employeeId}`, {
+            method: "DELETE",
+    })
+    .then(getEmployees)
+}
+
     return (
         <EmployeeContext.Provider value={{
-            employees, getEmployees, addEmployee
+            employees, getEmployees, addEmployee, removeEmployee
         }}>
             {props.children}
         </EmployeeContext.Provider>
